+++ patlas/db_manager/db_app/static/js/dropdowns/dropdownPopulation.js
@@ -45,22 +45,8 @@ getArrayTaxa().done((json) => {
   singleDropdownPopulate("#genusList2", listGenera, false)
   singleDropdownPopulate("#speciesList2", listSpecies, false)
 
-  // clickable <li> and control of displayer of current filters
-  const classArray = [".OrderClass", ".FamilyClass", ".GenusClass", ".SpeciesClass"]
-  for (let i = 0; i < classArray.length; i++) {
-    $(classArray[i]).on("click", function() {
-      // fill panel group displaying current selected taxa filters //
-      const stringClass = this.className.slice(0, -5)
-      const tempVar = this.firstChild.innerHTML
-
-      // checks if a taxon is already in display
-      const divStringClass = "#p_" + stringClass
-
-      filterDisplayer(tempVar, stringClass, divStringClass)
-    })
-  }
 })
-// }
+
 
 /**
  * Part of the code that populates the resistance dropdowns available through
@@ -89,35 +75,19 @@ getArrayRes().done( (json) => {
 
   // populate the menus for resistance filters
   singleDropdownPopulate("#cardList", listCard, "CardClass")
-  singleDropdownPopulate("#resList", listRes, "ResfinderClass")
+  singleDropdownPopulate("#resfinderList", listRes, "ResfinderClass")
 
   // populate the menus for intercection filters
   singleDropdownPopulate("#resCardList2", listCard, false)
   singleDropdownPopulate("#resResfinderList2", listRes, false)
 
-
-  const classArray = [".CardClass", ".ResfinderClass"]
-  for (let i = 0; i < classArray.length; i++) {
-    $(classArray[i]).on("click", function() {
-      // fill panel group displaying current selected taxa filters //
-      const stringClass = this.className.slice(0, -5)
-      const tempVar = this.firstChild.innerHTML
-
-      // checks if a taxon is already in display
-      const divStringClass = "#p_" + stringClass
-
-      filterDisplayer(tempVar, stringClass, divStringClass)
-    })
-  }
 })
-// }
 
 
 /**
  * Code that gests plasmidfinder dropdowns populated. Both through browse -->
  * Plasmid Families and browse --> advanced multiple
  */
-// if (firstInstace === true && pageReload === false) {
 getArrayPf().done((json) => {
   // first parse the json input file
   const listPF = []
@@ -134,29 +104,18 @@ getArrayPf().done((json) => {
   })
 
   // populate the menus for plasmid finder filter
-  singleDropdownPopulate("#plasmidFamiliesList", listPF, "PlasmidfinderClass")
+  singleDropdownPopulate("#plasmidFinderList", listPF, "PlasmidFinderClass")
 
   // populate the menus for the intercection filters
   singleDropdownPopulate("#pfList2", listPF, false)
 
-  $(".PlasmidfinderClass").on("click", function() {
-    // fill panel group displaying current selected taxa filters //
-    const stringClass = this.className.slice(0, -5)
-    const tempVar = this.firstChild.innerHTML
-    // checks if a taxon is already in display
-    const divStringClass = "#p_" + stringClass
-
-    filterDisplayer(tempVar, stringClass, divStringClass)
-  })
 })
-// }
 
 
 /**
  * Code that gets the virulence dropdowns populated, both through browse -->
  * virulence or browse --> advanced multiple
  */
-// if (firstInstace === true && pageReload === false) {
 getArrayVir().done( (json) => {
   // first parse the json input file
   const listVir = []
@@ -173,19 +132,9 @@ getArrayVir().done( (json) => {
   })
 
   // populate the menus virulence filters
-  singleDropdownPopulate("#virList", listVir, "VirulenceClass")
+  singleDropdownPopulate("#virulenceList", listVir, "VirulenceClass")
 
   // populate the menus for the intercection filters
   singleDropdownPopulate("#virList2", listVir, false)
 
-  $(".VirulenceClass").on("click", function() {
-    // fill panel group displaying current selected taxa filters //
-    const stringClass = this.className.slice(0, -5)
-    const tempVar = this.firstChild.innerHTML
-    // checks if a taxon is already in display
-    const divStringClass = "#p_" + stringClass
-
-    filterDisplayer(tempVar, stringClass, divStringClass)
-  })
 })
-// }