+++ patlas/db_manager/db_app/static/js/node_handling/node_removal_taxa.js
@@ -232,11 +232,11 @@ const requesterDB = (g, listGiFilter, counter, renderGraph, graphics,
 
           $("#virSubmit").click()
 
-        } else if (selectedFilter === "intersect") {
+        } else if (selectedFilter === "∩") {
 
           $("#intersectionsModalSubmit").click()
 
-        } else if (selectedFilter === "union") {
+        } else if (selectedFilter === "∪") {
 
           $("#unionModalSubmit").click()
 
@@ -324,7 +324,7 @@ const actualRemoval = (g, graphics, onload, forgetListGiFilter) => {
 const showDiv = () => {
   return new Promise( (resolve) => {
     // disables this button group
-    $("#toolButtonGroup button").attr("disabled", "disabled")
+    // $("#toolButtonGroup button").attr("disabled", "disabled")
     resolve($("#loading").show())
   })
 }
@@ -336,7 +336,7 @@ const showDiv = () => {
  */
 const hideDivsFileInputs = () => {
   $("#loading").hide()
-  $("#toolButtonGroup button").removeAttr("disabled")
+  // $("#toolButtonGroup button").removeAttr("disabled")
   $("#slideRight").prop("disabled", false)
   $("#slideLeft").prop("disabled", false)
 }