+++ patlas/db_manager/db_app/views.py
@@ -39,11 +39,11 @@ def index():
 
 @app.route("/test")
 def main_summary():
-    return repetitiveFunction("db_app/static/json/import_to_vivagraph.json")
+    return repetitiveFunction("db_app/static/json/import_to_vivagraph_v1.4.1.json")
 
 @app.route("/fullDS")
 def full_ds():
-    return repetitiveFunction("db_app/static/json/filtered_19012018.json")
+    return repetitiveFunction("db_app/static/json/filtered_v1.4.1.json")
 
 @app.route("/taxa")
 def taxa_summary():